Move join type input swapping to pub methods on Joins
